Move CodeNode limit resolution into limits module
